♻️ Refactor before adding subcommands
Refactors the existing implementation in advance of adding additional
subcommands.

dist/main.js: 3.3kb (+3%)
