+++ fixtures/fixtures-root.ts
@@ -1,0 +1,1 @@
+export const fixturesRoot = import.meta.dirname;
