+++ src/version.ts
@@ -1,0 +1,2 @@
+declare let VERSION: string; // Defined by esbuild
+export const version = VERSION;
